为一些常量加入强转

+++ src/core/raiden.h
@@ -104,14 +104,14 @@ static constexpr Float MaxFloat = std::numeric_limits<Float>::max();
 #define RAIDEN_NAMESPACE_END   };	
 
 
-static constexpr Float Pi = 3.14159265358979323846;
-static constexpr Float InvPi = 0.31830988618379067154;
-static constexpr Float Inv2Pi = 0.15915494309189533577;
-static constexpr Float Inv4Pi = 0.07957747154594766788;
-static constexpr Float PiOver2 = 1.57079632679489661923;
-static constexpr Float PiOver4 = 0.78539816339744830961;
-static constexpr Float Sqrt2 = 1.41421356237309504880;
-static constexpr Float ShadowEpsilon = 0.0001f;
+static constexpr Float Pi = (Float)3.14159265358979323846;
+static constexpr Float InvPi = (Float)0.31830988618379067154;
+static constexpr Float Inv2Pi = (Float)0.15915494309189533577;
+static constexpr Float Inv4Pi = (Float)0.07957747154594766788;
+static constexpr Float PiOver2 = (Float)1.57079632679489661923;
+static constexpr Float PiOver4 = (Float)0.78539816339744830961;
+static constexpr Float Sqrt2 = (Float)1.41421356237309504880;
+static constexpr Float ShadowEpsilon = (Float)0.0001f;
 
 template<typename T> class Vector3;
 template<typename T> class Vector2;
